requirements.txt
streamlit
pandas
numpy
streamlit_lottie
st_pages
streamlit.extras
gspread
google-api-python-client